+++ src/containerapp/HISTORY.rst
@@ -5,11 +5,16 @@ Release History
 upcoming
 ++++++
 * 'az containerapp function show/list': New command group to list and show functions in container apps.
+* 'az containerapp function list-keys': List function keys for a specific function in a container app
+* 'az containerapp function update-keys': Update specific function key for a specific function in a container app
+* 'az containerapp function list-hostkeys': List host keys for a container app
+* 'az containerapp function update-hostkeys': Update specific host key for a container app
 * 'az containerapp update/up': Disallow changing `--revisions-mode` to Labels.
 * 'az containerapp session code-interpreter': Fix `--path` in examples
 * 'az containerapp sessionpool create/update': Support `--lifecycle-type` and `--max-alive-period`
 * 'az containerapp up': support for `--kind` parameter
 * 'az containerapp env premium-ingress': Deprecate `--min-replicas` and `--max-replicas` parameters, use workload profile scale instead.
+* 'az containerapp sessionpool create/update': Support `--probe-yaml`
 
 1.2.0b3
 ++++++